Move txmgr in evm package
